Make code generation configurable, support OpenAPI
The first change is to make the code generation configurable, to support
different CRDs.

Also start supporting the full OpenAPI, including references, to allow
building against the core API (so far, only 'apps/v1' is tested). This
seems to be OK for most things, but we'll have to special-case
ObjectMeta.