Convert all types to DynamicValue
This effectively side-steps the entire Terraform type-system. The issue
is that in Terraform, a list containing `DynamicValue`, still has to
have all of the dynamic values reconcilable to the same type. This
doesn't apply to `x-kubernetes-int-or-string`, though there it can be
worked around by converting all to string. However, with more
complicated types that are `x-kubernetes-preserve-unknown-fields`, this
will not be possible.

Therefore, make everything dynamic, and handle checking ourselves. The
one downside is that IDE integrations etc can't use the schema types for
type-hints.

This may be worth revisiting when github.com/hashicorp/terraform#33916
is resolved.